updated app.vue to just import and use components
